requirements.txt
ccxt==4.3.50
pandas==2.2.2
pandas-ta-classic==0.3.14b1
psycopg2-binary==2.9.9
requests==2.32.3
hurst==0.0.5
scipy==1.13.1
tenacity==8.3.0
numpy==1.26.4
numba==0.61.2
fastapi==0.111.0
uvicorn[standard]==0.29.0
tqdm==4.66.4
pymongo==4.8.0
uvicorn==0.38.0
fastapi==0.120.0
hurst==0.0.5
pymongo==4.15.3
tenacity==9.1.2
dotenv==0.9.9
python-dotenv==1.2.1